Move policy evaluation as well
