+++ app/sidekiq/execute_task_stage_job.rb
@@ -7,7 +7,7 @@ class ExecuteTaskStageJob
     @task_stage = TaskStage.find(args.first)
     @task_stage.update(status: 'running')
     @task_stage.policy_evaluations.each do |policy_evaluation|
-      PolicyEvaluation::RunPolicyEvaluationJob.perform_async(policy_evaluation.id)
+      RunPolicyEvaluationJob.perform_async(policy_evaluation.id)
     end
     @task_stage.task_results.each do |task_result|
       task_result.update(status: 'running')
