--- app/sidekiq/policy_evaluation/policy_evaluation_failed_job.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module PolicyEvaluation
-  class PolicyEvaluationFailedJob
-    include Sidekiq::Job
-
-    def perform(*args); end
-  end
-end
