--- app/sidekiq/policy_evaluation/policy_evaluation_passed_job.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module PolicyEvaluation
-  class PolicyEvaluationPassedJob
-    include Sidekiq::Job
-
-    def perform(*args)
-      @policy_evaluation = PolicyEvaluation.find(args.first)
-    end
-  end
-end
