--- app/sidekiq/policy_evaluation/run_policy_evaluation_job.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module PolicyEvaluation
-  class RunPolicyEvaluationJob
-    include Sidekiq::Job
-
-    def perform(*args); end
-  end
-end
