+++ app/sidekiq/policy_evaluation_errored_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class PolicyEvaluationErroredJob
+  include Sidekiq::Job
+
+  def perform(*args); end
+end
