+++ app/sidekiq/policy_evaluation_overridden_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class PolicyEvaluationOverriddenJob
+  include Sidekiq::Job
+
+  def perform(*args); end
+end
