+++ app/sidekiq/policy_evaluation_passed_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class PolicyEvaluationPassedJob
+  include Sidekiq::Job
+
+  def perform(*args)
+    @policy_evaluation = PolicyEvaluation.find(args.first)
+  end
+end
