+++ app/sidekiq/run_policy_evaluation_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class RunPolicyEvaluationJob
+  include Sidekiq::Job
+
+  def perform(*args); end
+end
